requirements.txt
supabase
openai
google-genai
loguru
python-dotenv~=1.0.1
